requirements.txt
requests
pydantic
tenacity
python-dotenv